+++ src/main/java/ua/rent/masters/easystay/controller/BookingController.java
@@ -19,9 +19,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import ua.rent.masters.easystay.dto.request.BookingRequestDto;
-import ua.rent.masters.easystay.dto.request.BookingRequestUpdateDto;
-import ua.rent.masters.easystay.dto.response.BookingResponseDto;
+import ua.rent.masters.easystay.dto.booking.BookingRequestDto;
+import ua.rent.masters.easystay.dto.booking.BookingRequestUpdateDto;
+import ua.rent.masters.easystay.dto.booking.BookingResponseDto;
 import ua.rent.masters.easystay.model.BookingStatus;
 import ua.rent.masters.easystay.model.User;
 import ua.rent.masters.easystay.service.BookingService;
