+++ src/main/java/ua/rent/masters/easystay/dto/booking/BookingRequestUpdateDto.java
@@ -1,19 +1,15 @@
-package ua.rent.masters.easystay.dto.request;
+package ua.rent.masters.easystay.dto.booking;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 
-public record BookingRequestDto(
+public record BookingRequestUpdateDto(
         @NotNull
         LocalDate checkInDate,
         @NotNull
         LocalDate checkOutDate,
-        @NotNull
         @Min(value = 1, message = "The value must be at least 1")
-        Long accommodationId,
-        @NotNull
-        @Min(value = 1, message = "The value must be at least 1")
-        Long userId
+        Long accommodationId
 ) {
 }
