+++ src/main/java/ua/rent/masters/easystay/dto/booking/BookingResponseDto.java
@@ -1,9 +1,10 @@
-package ua.rent.masters.easystay.dto.response;
+package ua.rent.masters.easystay.dto.booking;
 
 import java.time.LocalDate;
 import ua.rent.masters.easystay.model.BookingStatus;
 
 public record BookingResponseDto(
+        Long id,
         LocalDate checkInDate,
         LocalDate checkOutDate,
         Long accommodationId,
