+++ src/main/java/ua/rent/masters/easystay/dto/payment/PaymentCancelResponseDto.java
@@ -1,4 +1,4 @@
-package ua.rent.masters.easystay.dto;
+package ua.rent.masters.easystay.dto.payment;
 
 public record PaymentCancelResponseDto(
         String content
