+++ src/main/java/ua/rent/masters/easystay/dto/payment/PaymentResponseDto.java
@@ -1,4 +1,4 @@
-package ua.rent.masters.easystay.dto;
+package ua.rent.masters.easystay.dto.payment;
 
 import java.math.BigDecimal;
