+++ src/main/java/ua/rent/masters/easystay/mapper/BookingMapper.java
@@ -2,8 +2,8 @@ package ua.rent.masters.easystay.mapper;
 
 import org.mapstruct.Mapper;
 import ua.rent.masters.easystay.config.MapperConfig;
-import ua.rent.masters.easystay.dto.request.BookingRequestDto;
-import ua.rent.masters.easystay.dto.response.BookingResponseDto;
+import ua.rent.masters.easystay.dto.booking.BookingRequestDto;
+import ua.rent.masters.easystay.dto.booking.BookingResponseDto;
 import ua.rent.masters.easystay.model.Booking;
 
 @Mapper(config = MapperConfig.class)
