+++ src/main/java/ua/rent/masters/easystay/mapper/PaymentMapper.java
@@ -3,7 +3,7 @@ package ua.rent.masters.easystay.mapper;
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
 import ua.rent.masters.easystay.config.MapperConfig;
-import ua.rent.masters.easystay.dto.PaymentResponseDto;
+import ua.rent.masters.easystay.dto.payment.PaymentResponseDto;
 import ua.rent.masters.easystay.model.Payment;
 
 @Mapper(config = MapperConfig.class)
