+++ src/main/java/ua/rent/masters/easystay/service/BookingService.java
@@ -2,9 +2,9 @@ package ua.rent.masters.easystay.service;
 
 import java.util.List;
 import org.springframework.data.domain.Pageable;
-import ua.rent.masters.easystay.dto.request.BookingRequestDto;
-import ua.rent.masters.easystay.dto.request.BookingRequestUpdateDto;
-import ua.rent.masters.easystay.dto.response.BookingResponseDto;
+import ua.rent.masters.easystay.dto.booking.BookingRequestDto;
+import ua.rent.masters.easystay.dto.booking.BookingRequestUpdateDto;
+import ua.rent.masters.easystay.dto.booking.BookingResponseDto;
 import ua.rent.masters.easystay.model.BookingStatus;
 
 public interface BookingService {
