+++ src/main/java/ua/rent/masters/easystay/service/PaymentService.java
@@ -1,8 +1,8 @@
 package ua.rent.masters.easystay.service;
 
 import java.util.List;
-import ua.rent.masters.easystay.dto.PaymentCancelResponseDto;
-import ua.rent.masters.easystay.dto.PaymentResponseDto;
+import ua.rent.masters.easystay.dto.payment.PaymentCancelResponseDto;
+import ua.rent.masters.easystay.dto.payment.PaymentResponseDto;
 import ua.rent.masters.easystay.model.User;
 
 public interface PaymentService {
